Fix typos, address Rider refactoring feedback
