+++ src/SpikeCore/SpikeCore.MessageBus.Foundatio/AutofacIntegration/ContainerBuilderExtensions.cs
@@ -3,13 +3,11 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
-
 using Autofac;
 using Autofac.Core;
-
 using Foundatio.Messaging;
 
-namespace SpikeCore.MessageBus.Foundatio.AutofacIntergration
+namespace SpikeCore.MessageBus.Foundatio.AutofacIntegration
 {
     public static class ContainerBuilderExtensions
     {
@@ -17,7 +15,7 @@ namespace SpikeCore.MessageBus.Foundatio.AutofacIntergration
         {
             var messageBus = new InMemoryMessageBus() as IMessageBus;
 
-            // Register the message bus so it's still availible as a service if needs be.
+            // Register the message bus so it's still available as a service if needs be.
             self
                 .RegisterInstance(messageBus)
                 .As<IMessageBus>()
@@ -25,12 +23,11 @@ namespace SpikeCore.MessageBus.Foundatio.AutofacIntergration
 
             var subscribeAsyncMethodInfo = typeof(IMessageSubscriber)
                 .GetMethods()
-                .Where(methodInfo => methodInfo.Name == "SubscribeAsync")
-                .Single();
+                .Single(methodInfo => methodInfo.Name == "SubscribeAsync");
 
             self.RegisterBuildCallback(container =>
             {
-                // Find the registerd types which impliment IMessageHandler<>
+                // Find the registered types which implement IMessageHandler<>
                 var messageHandlerRegistrations = container
                     .ComponentRegistry
                     .Registrations
@@ -39,7 +36,7 @@ namespace SpikeCore.MessageBus.Foundatio.AutofacIntergration
 
                 foreach (var messageHandlerRegistration in messageHandlerRegistrations)
                 {
-                    // Find all the individual IMessageHandler<> implimentations on the type
+                    // Find all the individual IMessageHandler<> implementations on the type
                     var iMessageHandlerTypes = messageHandlerRegistration
                         .Activator
                         .LimitType
@@ -47,7 +44,7 @@ namespace SpikeCore.MessageBus.Foundatio.AutofacIntergration
                         .Where(interfaceType => IsMessageHandler(interfaceType))
                         .ToList();
 
-                    if (iMessageHandlerTypes.Count() > 0)
+                    if (iMessageHandlerTypes.Any())
                     {
                         // When the instance has been created
                         messageHandlerRegistration.Activated += (sender, eventArgs) =>
@@ -78,8 +75,8 @@ namespace SpikeCore.MessageBus.Foundatio.AutofacIntergration
                                 // Call SubscribeAsync passing in the subscribeHandler
                                 var task = (Task)subscribeAsyncMethodInfo
                                     .MakeGenericMethod(messageType)
-                                    // Type.Missing is used to reprisent using the default value of an optional parameter
-                                    .Invoke(messageBus, new object[] { subscribeHandler, Type.Missing });
+                                    // Type.Missing is used to represent using the default value of an optional parameter
+                                    .Invoke(messageBus, new[] { subscribeHandler, Type.Missing });
 
                                 // This might not be needed. It might be fine to have the subscribe carry on asynchronously
                                 task.Wait();
